refactor: Unify call-actor tool across default and apps modes
Single definition in default/call_actor.ts; the two mode variants differ
only in description (apps appends the widget addendum). Deletes the
duplicated apps/call_actor.ts.

Fixes #975
